Histogram enhancements, -f added to Fast(rm,mv,cp)
HIstograms can be either of the # of unique k-mers U(f) with each frequency f, or the # of k-mer instances with a given frequency I(f).  I(f) = f*U(f) except at the boundaries if the range of the histogram is truncated.  Code and data format change to handle and support this distinction.