New release 1.1: More space efficient tables, Fast(mv|cp) upgrade, Nremover & Ncontiger added
Tables now have a prefix table over a fixed # of bytes, that are then no longer stored in each entry saving a great deal of space at the expense of having to determine the prefix for a given index.  Changes through system, especially the library and FastK

Both Fastcp and Fastmv can now move multiple sources when aa directory is the last argument.  Thiss was an oversight as cp and mv do this and the idea is that Fastcp and Fastmv mimic the action of the commands on FastK outputs.

Prefilters Ncontiger and Nremover are introduced to remove N's in a meaningful way for both scaffolds of contigs (Ncontiger) and low quality bases (Nremover).